+++ backend/backend/urls.py
@@ -26,8 +26,7 @@ from rest_framework.settings import api_settings
 from backend.views import info_view
 from backend.views import obtain_auth_token
 from node.urls import router as node_router
-from substrapp.urls import compute_plan_router
-from substrapp.urls import router
+from substrapp.urls import urlpatterns
 from users.urls import router as user_router
 
 urlpatterns = (
@@ -36,7 +35,7 @@ urlpatterns = (
             r"",
             include(
                 [
-                    path("", include((router.urls + compute_plan_router.urls, "substrapp"))),
+                    path("", include((urlpatterns, "substrapp"))),
                     path("", include((node_router.urls, "node"))),
                     path("", include((user_router.urls, "user"))),  # for secure jwt authent
                     path("api-token-auth/", obtain_auth_token),  # for expiry token authent