+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Hardcode task outputs (#1068)
 - Cross asset filtering by key on asset views (#1070)
 - Health service for the event app (#1089)
+- (BREAKING) New register tasks endpoint (#1053)
 
 ### Changed